// src/store/user/types.ts
export const enum UserActionTypes {
	GET_USER_SUCCESS = 'GET_USER_SUCCESS',
	GET_USER_ERROR = 'GET_USER_ERROR',
	REMOVE_USER = 'REMOVE_USER',
	CREATE_USER = 'CREATE_USER',
	CREATE_USER_ERROR = 'CREATE_USER_ERROR',
	CLEAR_ERRORS = 'CLEAR_ERRORS',
	SET_USER = 'SET_USER',
	SET_TOKEN = 'SET_TOKEN',
	LOGOUT = 'LOGOUT',
	LOGOUT_SUCCESS = 'LOGOUT_SUCCESS',
}